Worked on choose character and <teamname_url>
I finished choose character but didn't test it. 
I worked on <teamname_url> because I wanted the page to detect when the characters are chosen or not. This is working except for validate host. The host is false in the html. 
Maybe join the 3 tables instead of this.. Better design? I don't know, when the characters aren't assigned it doesn't have to query the other tables.